requirements.txt
fastapi
uvicorn[standard]
jinja2
pymysql
python-dotenv